Update pkgrepo.absent to not use named argument when calling del_repo
nipkg.py module is using alias as the parameter name. Because of this,
pkgrepo.absent is failing when applying a state. Instead of updating
nipkg.py to use repo instead of alias and be consistent with salt API
which is a breaking change that could possibly affect clients,
we are dropping the named argument since it is used only in this
function so there is no real value in having it.

This is probably some old code that nobody touched in the last years.

Signed-off-by: Cristian Hotea <cristian.hotea@ni.com>